Resolving issue 15
Issue: https://cpplinq.codeplex.com/workitem/15

The error was subtle but the fix is simple, first mistakenly
used the previous range return_type.  This is not correct for
builders as it might be a reference. Instead use the previous
range value_type which is never a reference.

In certain situations this caused the returned value to be
freed before the outer scope could copy it.

Thanks to Sepidar for finding the issue and posting a repro.
The repro is now added as a test case.